+++ bin/commands/components/install/extract/index.ts
@@ -150,6 +150,20 @@ export function execute(componentFile: string, autoEncoding?: string, upgrade?: 
     common.printErrorAndExit(`Error ZWEL0167E: Cannot find component name from ${componentFile} package manifest`, undefined, 167);
   }
   common.printDebug(`- Component name found as ${componentName}`);
+
+  // If the component has appfw plugins, their validity should be checked against appfw plugin schema.
+  //   If invalid, the installation will exit with an error message.
+  if (manifest.appfwPlugins) {
+    manifest.appfwPlugins.forEach((appfwPlugin: {path: string})=> {
+      let result = component.getPluginDefinition(pathoid.resolve(tmpDir, appfwPlugin.path), true);
+      //Normally, getPluginDefinition would quit upon failure. But we want to cleanup the tmpDir before that.
+      // So, we pass true to allow it to continue, check for null, and then remove the tmpdir and exit if so.
+      if (result === null) {
+        fs.rmrf(tmpDir);
+        std.exit(1);
+      }
+    });
+  }
   
   const destinationDir = pathoid.resolve(targetDir, componentName);
   const bkpDir = pathoid.resolve(targetDir, `${componentName}_zwebkp`);